+++ public/index.html
@@ -125,7 +125,7 @@
       
         <li><a href="formateo-de-cadenas">Formateo de cadenas</a></li>
       
-        <li><a href="regular-expressions">Regular Expressions</a></li>
+        <li><a href="expresiones-regulares">Expresiones regulares</a></li>
       
         <li><a href="json">JSON</a></li>
